requirements.txt
numpy
rasterio
tqdm
pillow
imagehash
